cleaned up the save and return per page code
